Reorganize tests by platform directory
- Move Linux-specific tests to test/linux/ directory
- Update import paths from ../test_helper.dart to ../../test_helper.dart
- Remove @TestOn('linux') annotations
- Update CI scripts to run tests from appropriate directories
- ci-test-common.sh runs test/ excluding test/linux/
- ci-test-linux.sh runs only test/linux/
- Remove empty test/processors/ directory
