+++ test/linux/crop_comparison_test.dart
@@ -4,7 +4,6 @@ import 'dart:math' as math;
 import 'dart:ui' as ui;
 import 'package:flutter_test/flutter_test.dart';
 
-@TestOn('linux')
 import 'package:aks/services/processors/cpu_processor.dart';
 import 'package:aks/services/processors/vulkan_processor.dart';
 import 'package:aks/services/processors/vulkan/vulkan_bindings.dart';
@@ -13,7 +12,7 @@ import 'package:aks/models/adjustments.dart';
 import 'package:aks/models/crop_state.dart';
 import 'package:aks/services/raw_processor.dart';
 import 'package:aks/services/image_processor.dart';
-import '../test_helper.dart';
+import '../../test_helper.dart';
 
 void main() {
   group('CPU vs GPU Crop Comparison Tests', () {
