+++ test/linux/debug_crop_test.dart
@@ -1,13 +1,12 @@
 import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 
-@TestOn('linux')
 import 'package:aks/services/processors/vulkan_processor.dart';
 import 'package:aks/services/processors/vulkan/vulkan_bindings.dart';
 import 'package:aks/models/crop_state.dart';
 import 'package:aks/services/processors/image_processor_interface.dart';
 import 'package:aks/services/raw_processor.dart';
-import '../test_helper.dart';
+import '../../test_helper.dart';
 
 void main() {
   test('Debug crop calculations', () async {
