+++ test/linux/processor_comparison_test.dart
@@ -3,7 +3,6 @@ import 'dart:typed_data';
 import 'dart:math' as math;
 import 'package:flutter_test/flutter_test.dart';
 
-@TestOn('linux')
 import 'package:aks/services/processors/cpu_processor.dart';
 import 'package:aks/services/processors/vulkan_processor.dart';
 import 'package:aks/services/processors/vulkan/vulkan_bindings.dart';
@@ -13,7 +12,7 @@ import 'package:aks/models/edit_pipeline.dart';
 import 'package:aks/models/crop_state.dart';
 import 'package:aks/services/raw_processor.dart';
 import 'package:aks/services/image_processor.dart';
-import '../test_helper.dart';
+import '../../test_helper.dart';
 
 void main() {
   group('Processor Comparison Tests', () {
